feat(v1): added new assert function `assertSentCount` - Updated testing docs and reworked asserts presentation - fixed in `Response` that arguments body could be null
