+++ docs/commands/overview.md
@@ -1,5 +1,7 @@
 # Overview of Commands
 
+... better docs will come ;-)
+
 ## Requirements
 
 You need to install following package to use the code generation feature
