+++ src/Client/Request.php
@@ -63,7 +63,7 @@ class Request
     public function arguments(): array
     {
         $xml = SoapXml::fromString($this->xmlContent());
-
-        return Arr::first(XMLSerializer::domNodeToArray($xml->getBody()));
+        $arguments = Arr::first(XMLSerializer::domNodeToArray($xml->getBody()));
+        return $arguments ?? [];
     }
 }
